+++ src/main/java/com/jaramgroupware/mms/service/RoleService.java
@@ -30,7 +30,7 @@ public class RoleService {
      * @return 조회된 Role(Object)의 정보를 담은 dto, 해당 데이터가 없을 시 INVALID_ROLE_ID 예외 처리
      */
     @Transactional(readOnly = true)
-    public RoleResponseDto findById(Integer id){
+    public RoleResponseDto findById(Long id){
         Role targetRole = roleRepository.findById(id)
                 .orElseThrow(() -> new ServiceException(NOT_FOUND,"해당 Role을 찾을 수 없습니다."));
         return new RoleResponseDto(targetRole);
